+ ejs и убрал папку src
