+++ entities/book.js
