+++ package.json
@@ -19,6 +19,7 @@
   "homepage": "https://github.com/nikita-o/express-learn#readme",
   "dependencies": {
     "dotenv": "^16.0.1",
+    "ejs": "^3.1.8",
     "express": "^4.18.1",
     "multer": "^1.4.5-lts.1",
     "uuid": "^8.3.2"
