+++ routes/render/index.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res.render('index', {
+        title: 'start',
+    })
+});
+
+module.exports = router;
