+++ views/index.ejs
@@ -1,0 +1,5 @@
+<%- include('./layouts/header.ejs', {title: title}) %>
+
+<h1>Hello, world!</h1>
+
+<%- include('./layouts/footer.ejs') %>
